App/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
PyMySQL==1.0.2
SQLAlchemy==1.4.46
Werkzeug==2.2.2
